Section config guards and CampComplianceAccess leave Shell (nobodies-collective/Humans#1091)
- GoogleWorkspaceSettings binding + production credentials guard move into
  GoogleIntegration's Section.Register (HostDefaults.EnvironmentKey fail-closed
  pattern); its only reader already lives there
- EmailSettings, GoogleWorkspaceOptions, TicketVendorSettings stay in Shell —
  each has readers in multiple sections
- CampComplianceAccess policy, requirement and handler all move to Humans.Camps,
  the policy's consumer: policy registers in Camps' SectionPolicies, handler in
  its Section, requirement goes internal. The coordinator lookup remains a
  cross-section read through Shifts' contracts leaf. Registered from the
  consumer, the policy deactivates with its pages; Shifts' public surface
  shrinks by the requirement type
