+++ src/Sections/Humans.Camps/Authorization/CampComplianceAccessRequirement.cs
@@ -1,7 +1,7 @@
 using Humans.Base.Authorization;
 using Microsoft.AspNetCore.Authorization;
 
-namespace Humans.Shifts.Contracts;
+namespace Humans.Camps.Authorization;
 
 /// <summary>
 /// Succeeds when the user is a CampAdmin or Admin, OR is a coordinator /
@@ -9,8 +9,8 @@ namespace Humans.Shifts.Contracts;
 /// compliance matrix (<see cref="PolicyNames.CampComplianceAccess"/>) so camp
 /// admins and team coordinators can both see role staffing across barrios,
 /// without widening the CampAdmin-only management surface in
-/// <c>CampAdminController</c>. Lives under Contracts/ (unlike the internal
-/// handler) because Shell's AuthorizationPolicyExtensions constructs it
-/// directly to back the policy.
+/// <c>CampAdminController</c>. Moved from Shifts' Contracts/ folder with the
+/// handler at nobodies-collective/Humans#1091 — this section is the policy's
+/// only consumer, so nothing needs it public any more.
 /// </summary>
-public class CampComplianceAccessRequirement : IAuthorizationRequirement;
+internal sealed class CampComplianceAccessRequirement : IAuthorizationRequirement;
