+++ src/Sections/Humans.Camps/SectionPolicies.cs
@@ -1,6 +1,7 @@
 using Humans.Base.Authorization;
 using Humans.Base.Constants;
 using Humans.Base.Interfaces;
+using Humans.Camps.Authorization;
 using Microsoft.AspNetCore.Authorization;
 
 namespace Humans.Camps;
@@ -10,8 +11,13 @@ namespace Humans.Camps;
 /// alongside <see cref="Section"/> — nothing names it.
 /// </summary>
 /// <remarks>
-/// <c>CampComplianceAccess</c> stays in Shell: it also admits any team/sub-team coordinator
-/// (a Shifts-domain lookup), so it is a composite spanning two sections' roles.
+/// <c>CampComplianceAccess</c> registers here — not in Shifts, whose coordinator lookup
+/// its handler reads — because this section is the policy's consumer
+/// (<c>CampComplianceController</c>, the Compliance nav entry) and the derived
+/// section-dependency graph can't see a policy-name reference: a Shifts-owned
+/// registration could leave Camps 500ing "policy not found" if Shifts were ever
+/// deactivated alone. Registered from Camps, the policy leaves with its consumers
+/// (nobodies-collective/Humans#1091).
 /// </remarks>
 internal sealed class SectionPolicies : ISectionPolicies
 {
@@ -19,5 +25,11 @@ internal sealed class SectionPolicies : ISectionPolicies
     {
         options.AddPolicy(PolicyNames.CampAdminOrAdmin, policy =>
             policy.RequireRole(RoleNames.CampAdmin, RoleNames.Admin));
+
+        // CampAdmin/Admin OR any team coordinator — the OR (including the
+        // team-coordinator lookup) lives in Shifts' CampComplianceAccessHandler so the
+        // policy is a single requirement (policy requirements AND together).
+        options.AddPolicy(PolicyNames.CampComplianceAccess, policy =>
+            policy.AddRequirements(new CampComplianceAccessRequirement()));
     }
 }
